refactor: migrated setting-related methods from aiprovider to aiprovidersettings
