+++ src/SmartHopper.Providers.MistralAI/MistralAIProvider.cs
@@ -62,91 +62,6 @@ namespace SmartHopper.Providers.MistralAI
             }
         }
 
-        public override IEnumerable<SettingDescriptor> GetSettingDescriptors()
-        {
-            return new[]
-            {
-                new SettingDescriptor
-                {
-                    Name = "ApiKey",
-                    Type = typeof(string),
-                    DefaultValue = string.Empty,
-                    IsSecret = true,
-                    DisplayName = "API Key",
-                    Description = "Your MistralAI API key",
-                },
-                new SettingDescriptor
-                {
-                    Name = "Model",
-                    Type = typeof(string),
-                    DefaultValue = DefaultModelValue,
-                    IsSecret = false,
-                    DisplayName = "Model",
-                    Description = "The model to use for completions",
-                },
-                new SettingDescriptor
-                {
-                    Name = "MaxTokens",
-                    Type = typeof(int),
-                    DefaultValue = 150,
-                    IsSecret = false,
-                    DisplayName = "Max Tokens",
-                    Description = "Maximum number of tokens to generate",
-                },
-            };
-        }
-
-        public override bool ValidateSettings(Dictionary<string, object> settings)
-        {
-            Debug.WriteLine($"[MistralAI] ValidateSettings called. Settings null? {settings == null}");
-
-            if (settings == null)
-            {
-                return false;
-            }
-
-            // Extract values from settings dictionary
-            string apiKey = null;
-            string model = null;
-            int? maxTokens = null;
-
-            // Get API key if present
-            if (settings.TryGetValue("ApiKey", out var apiKeyObj) && apiKeyObj != null)
-            {
-                apiKey = apiKeyObj.ToString();
-                Debug.WriteLine($"[MistralAI] API key extracted (length: {apiKey.Length})");
-            }
-
-            // Get model if present
-            if (settings.TryGetValue("Model", out var modelObj) && modelObj != null)
-            {
-                model = modelObj.ToString();
-                Debug.WriteLine($"[MistralAI] Model extracted: {model}");
-            }
-
-            // Check max tokens if present - must be a positive number
-            if (settings.TryGetValue("MaxTokens", out var maxTokensObj) && maxTokensObj != null)
-            {
-                // Try to parse as integer
-                if (int.TryParse(maxTokensObj.ToString(), out int parsedMaxTokens))
-                {
-                    maxTokens = parsedMaxTokens;
-                }
-            }
-            
-            // Use the centralized validation method for the common settings
-            bool isValid = true;
-            
-            // Only validate settings that are actually provided (partial updates allowed)
-            if (apiKey != null || model != null || maxTokens.HasValue)
-            {
-                isValid = MistralAIProviderSettings.ValidateSettingsLogic(apiKey, model, maxTokens);
-            }
-
-            Debug.WriteLine($"[MistralAI] Settings validation result: {isValid}");
-            return isValid;
-        }
-
         public override async Task<AIResponse> GetResponse(JArray messages, string model, string jsonSchema = "", string endpoint = "", bool includeToolDefinitions = false)
         {
             // Get settings from the secure settings store
