+++ src/SmartHopper.Providers.OpenAI/OpenAIProvider.cs
@@ -64,108 +64,6 @@ namespace SmartHopper.Providers.OpenAI
             }
         }
 
-        public override IEnumerable<SettingDescriptor> GetSettingDescriptors()
-        {
-            return new[]
-            {
-                new SettingDescriptor
-                {
-                    Name = "ApiKey",
-                    Type = typeof(string),
-                    DefaultValue = string.Empty,
-                    IsSecret = true,
-                    DisplayName = "API Key",
-                    Description = "Your OpenAI API key",
-                },
-                new SettingDescriptor
-                {
-                    Name = "Model",
-                    Type = typeof(string),
-                    DefaultValue = DefaultModelValue,
-                    IsSecret = false,
-                    DisplayName = "Model",
-                    Description = "The model to use for completions",
-                },
-                new SettingDescriptor
-                {
-                    Name = "MaxTokens",
-                    Type = typeof(int),
-                    DefaultValue = 150,
-                    IsSecret = false,
-                    DisplayName = "Max Tokens",
-                    Description = "Maximum number of tokens to generate",
-                },
-                new SettingDescriptor
-                {
-                    Name = "ReasoningEffort",
-                    Type = typeof(string),
-                    DefaultValue = "medium",
-                    IsSecret = false,
-                    DisplayName = "Reasoning Effort",
-                    Description = "Level of reasoning effort for reasoning models (low, medium, or high)",
-                    AllowedValues = new[] { "low", "medium", "high" },
-                },
-            };
-        }
-
-        public override bool ValidateSettings(Dictionary<string, object> settings)
-        {
-            Debug.WriteLine($"[OpenAI] ValidateSettings called. Settings null? {settings == null}");
-            if (settings == null)
-            {
-                return false;
-            }
-
-            // Extract values from settings dictionary
-            string apiKey = null;
-            string model = null;
-            string reasoningEffort = null;
-            int? maxTokens = null;
-
-            // Get API key if present
-            if (settings.TryGetValue("ApiKey", out var apiKeyObj) && apiKeyObj != null)
-            {
-                apiKey = apiKeyObj.ToString();
-                Debug.WriteLine($"[OpenAI] API key extracted (length: {apiKey.Length})");
-            }
-
-            // Get model if present
-            if (settings.TryGetValue("Model", out var modelObj) && modelObj != null)
-            {
-                model = modelObj.ToString();
-                Debug.WriteLine($"[OpenAI] Model extracted: {model}");
-            }
-
-            // Get reasoning effort if present
-            if (settings.TryGetValue("ReasoningEffort", out var reasoningEffortObj) && reasoningEffortObj != null)
-            {
-                reasoningEffort = reasoningEffortObj.ToString();
-                Debug.WriteLine($"[OpenAI] ReasoningEffort extracted: {reasoningEffort}");
-            }
-
-            // Check max tokens if present - must be a positive number
-            if (settings.TryGetValue("MaxTokens", out var maxTokensObj) && maxTokensObj != null)
-            {
-                // Try to parse as integer
-                if (int.TryParse(maxTokensObj.ToString(), out int parsedMaxTokens))
-                {
-                    maxTokens = parsedMaxTokens;
-                }
-            }
-
-            // Use the centralizes validation method for the common settings
-            bool isValid = true;
-
-            // Only validate settings that are actually provided (partial updates allowed)
-            if (apiKey != null || model != null || maxTokens.HasValue || reasoningEffort != null)
-            {
-                isValid = OpenAIProviderSettings.ValidateSettingsLogic(apiKey, model, maxTokens.Value, reasoningEffort);
-            }
-
-            Debug.WriteLine($"[OpenAI] Settings validation result: {isValid}");
-            return isValid;
-        }
-
         /// <summary>
         /// Sends messages to the OpenAI Chat Completions endpoint, injecting a reasoning summary parameter when supported
         /// and wrapping any returned reasoning_summary in <think> tags before the actual content.
