+++ src/SmartHopper.Providers.OpenAI/OpenAIProviderSettings.cs
@@ -12,9 +12,9 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Windows.Forms;
 using SmartHopper.Config.Dialogs;
 using SmartHopper.Config.Interfaces;
+using SmartHopper.Config.Models;
 
 namespace SmartHopper.Providers.OpenAI
 {
@@ -23,13 +23,9 @@ namespace SmartHopper.Providers.OpenAI
     /// This class is responsible for creating the UI controls for configuring the provider
     /// and for managing the provider's settings.
     /// </summary>
-    public class OpenAIProviderSettings : AIProviderSettings, IDisposable
+    public class OpenAIProviderSettings : AIProviderSettings
     {
         private new readonly OpenAIProvider provider;
-        private new TextBox apiKeyTextBox;
-        private new TextBox modelTextBox;
-        private new NumericUpDown maxTokensNumeric;
-        private new ComboBox reasoningEffortComboBox;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="OpenAIProviderSettings"/> class.
@@ -41,18 +37,98 @@ namespace SmartHopper.Providers.OpenAI
             this.provider = provider ?? throw new ArgumentNullException(nameof(provider));
         }
 
-        /// <summary>
-        /// Internal method for validating OpenAI settings.
-        /// </summary>
-        /// <param name="apiKey">The API key to validate.</param>
-        /// <param name="model">The model name to validate.</param>
-        /// <param name="maxTokens">The max tokens setting to validate.</param>
-        /// <param name="reasoningEffort">The reasoning effort setting to validate.</param>
-        /// <param name="showErrorDialogs">Whether to show error dialogs for validation failures.</param>
-        /// <returns>True if all provided settings are valid, otherwise false.</returns>
-        internal static bool ValidateSettingsLogic(string apiKey, string model, int maxTokens, string reasoningEffort, bool showErrorDialogs = false)
+        public override IEnumerable<SettingDescriptor> GetSettingDescriptors()
         {
+            return new[]
+            {
+                new SettingDescriptor
+                {
+                    Name = "ApiKey",
+                    Type = typeof(string),
+                    DefaultValue = string.Empty,
+                    IsSecret = true,
+                    DisplayName = "API Key",
+                    Description = "Your OpenAI API key",
+                },
+                new SettingDescriptor
+                {
+                    Name = "Model",
+                    Type = typeof(string),
+                    DefaultValue = this.provider.DefaultModel,
+                    IsSecret = false,
+                    DisplayName = "Model",
+                    Description = "The model to use for completions",
+                },
+                new SettingDescriptor
+                {
+                    Name = "MaxTokens",
+                    Type = typeof(int),
+                    DefaultValue = 150,
+                    IsSecret = false,
+                    DisplayName = "Max Tokens",
+                    Description = "Maximum number of tokens to generate",
+                },
+                new SettingDescriptor
+                {
+                    Name = "ReasoningEffort",
+                    Type = typeof(string),
+                    DefaultValue = "medium",
+                    IsSecret = false,
+                    DisplayName = "Reasoning Effort",
+                    Description = "Level of reasoning effort for reasoning models (low, medium, or high)",
+                    AllowedValues = new[] { "low", "medium", "high" },
+                },
+            };
+        }
+
+        public override bool ValidateSettings(Dictionary<string, object> settings)
+        {
+            Debug.WriteLine($"[OpenAI] ValidateSettings called. Settings null? {settings == null}");
+            if (settings == null)
+            {
+                return false;
+            }
+
+            // Extract values from settings dictionary
+            string apiKey = null;
+            string model = null;
+            string reasoningEffort = null;
+            int? maxTokens = null;
+
+            // Get API key if present
+            if (settings.TryGetValue("ApiKey", out var apiKeyObj) && apiKeyObj != null)
+            {
+                apiKey = apiKeyObj.ToString();
+                Debug.WriteLine($"[OpenAI] API key extracted (length: {apiKey.Length})");
+            }
+
+            // Get model if present
+            if (settings.TryGetValue("Model", out var modelObj) && modelObj != null)
+            {
+                model = modelObj.ToString();
+                Debug.WriteLine($"[OpenAI] Model extracted: {model}");
+            }
+
+            // Get reasoning effort if present
+            if (settings.TryGetValue("ReasoningEffort", out var reasoningEffortObj) && reasoningEffortObj != null)
+            {
+                reasoningEffort = reasoningEffortObj.ToString();
+                Debug.WriteLine($"[OpenAI] ReasoningEffort extracted: {reasoningEffort}");
+            }
+
+            // Check max tokens if present - must be a positive number
+            if (settings.TryGetValue("MaxTokens", out var maxTokensObj) && maxTokensObj != null)
+            {
+                // Try to parse as integer
+                if (int.TryParse(maxTokensObj.ToString(), out int parsedMaxTokens))
+                {
+                    maxTokens = parsedMaxTokens;
+                }
+            }
+
             Debug.WriteLine($"Validating OpenAI settings: API Key: {apiKey}, Model: {model}, Max Tokens: {maxTokens}, Reasoning Effort: {reasoningEffort}");
+
+            var showErrorDialogs = true; // Set to false if you don't want to show error dialogs
 
             // Skip API key validation since any value is valid
 
@@ -67,7 +143,7 @@ namespace SmartHopper.Providers.OpenAI
                 }
                 return false;
             }
-            
+
             // Check if reasoning effort is valid
             if (string.IsNullOrWhiteSpace(reasoningEffort) || !new[] { "low", "medium", "high" }.Contains(reasoningEffort))
             {
@@ -79,11 +155,6 @@ namespace SmartHopper.Providers.OpenAI
             }
 
             return true;
-        }
-
-        public void Dispose()
-        {
-            throw new NotImplementedException();
         }
     }
 }
