+++ src/SmartHopper.Providers.Template/TemplateProvider.cs
@@ -90,96 +90,6 @@ namespace SmartHopper.Providers.Template
         }
 
         /// <summary>
-        /// Gets the setting descriptors for this provider.
-        /// These describe the settings that can be configured in the UI.
-        /// </summary>
-        /// <returns>A collection of setting descriptors.</returns>
-        public override IEnumerable<SettingDescriptor> GetSettingDescriptors()
-        {
-            // Define the settings that your provider requires
-            return new[]
-            {
-                new SettingDescriptor
-                {
-                    Name = "ApiKey",
-                    DisplayName = "API Key",
-                    Description = "Your API key for the Template service",
-                    IsSecret = true, // Set to true for sensitive data like API keys
-                    Type = typeof(string)
-                },
-                new SettingDescriptor
-                {
-                    Name = "Model",
-                    DisplayName = "Model",
-                    Description = "The model to use for generating responses",
-                    Type = typeof(string),
-                    DefaultValue = _defaultModel
-                },
-                new SettingDescriptor
-                {
-                    Name = "MaxTokens",
-                    DisplayName = "Max Tokens",
-                    Description = "Maximum number of tokens to generate",
-                    Type = typeof(int),
-                    DefaultValue = 150
-                }
-            };
-        }
-
-        /// <summary>
-        /// Validates the provided settings.
-        /// </summary>
-        /// <param name="settings">The settings to validate.</param>
-        /// <returns>True if the settings are valid, otherwise false.</returns>
-        public override bool ValidateSettings(Dictionary<string, object> settings)
-        {
-            // Only validate settings that are actually provided
-            if (settings == null)
-                return false;
-
-            // Extract values from settings dictionary
-            string apiKey = null;
-            string endpoint = null;
-            int? maxTokens = null;
-
-            // Get API key if present
-            if (settings.TryGetValue("ApiKey", out var apiKeyObj) && apiKeyObj != null)
-            {
-                apiKey = apiKeyObj.ToString();
-                Debug.WriteLine($"[TemplateProvider] API key extracted (length: {apiKey.Length})");
-            }
-
-            // Get endpoint if present
-            if (settings.TryGetValue("Endpoint", out var endpointObj) && endpointObj != null)
-            {
-                endpoint = endpointObj.ToString();
-                Debug.WriteLine($"[TemplateProvider] Endpoint extracted: {endpoint}");
-            }
-
-            // Check max tokens if present
-            if (settings.TryGetValue("MaxTokens", out var maxTokensObj) && maxTokensObj != null)
-            {
-                // Try to parse as integer
-                if (int.TryParse(maxTokensObj.ToString(), out int parsedMaxTokens))
-                {
-                    maxTokens = parsedMaxTokens;
-                    Debug.WriteLine($"[TemplateProvider] MaxTokens extracted: {maxTokens}");
-                }
-                else
-                {
-                    Debug.WriteLine($"[TemplateProvider] MaxTokens validation failed: not an integer, got {maxTokensObj}");
-                    return false;
-                }
-            }
-            
-            // Use the centralized validation method for the common settings
-            bool isValid = TemplateProviderSettings.ValidateSettingsLogic(apiKey, endpoint, maxTokens);
-            
-            Debug.WriteLine($"[TemplateProvider] Settings validation result: {isValid}");
-            return isValid;
-        }
-
-        /// <summary>
         /// Gets a response from the AI provider.
         /// </summary>
         /// <param name="messages">The messages to send to the AI provider.</param>
